requirements.txt
opencv-python
pymongo
python-dotenv
numpy
ultralytics
RPi.GPIO
pyserial
modbus_tk
requests
pygame
gpiozero